requirements.txt
openai
streamlit
Faker==0.7.4
python-dotenv